+++ src/plugins/data_source_management/public/components/data_source_menu/data_source_menu.tsx
@@ -15,19 +15,22 @@ import {
 import { MountPointPortal } from '../../../../opensearch_dashboards_react/public';
 import { DataSourceSelectable } from './data_source_selectable';
 import { DataSourceOption } from '../data_source_selector/data_source_selector';
+import { DataSourceView } from '../data_source_view';
 
 export interface DataSourceMenuProps {
-  showDataSourceSelectable: boolean;
+  showDataSourceSelectable?: boolean;
+  showDataSourceView?: boolean;
   appName: string;
-  savedObjects: SavedObjectsClientContract;
-  notifications: NotificationsStart;
+  savedObjects?: SavedObjectsClientContract;
+  notifications?: NotificationsStart;
   fullWidth: boolean;
-  hideLocalCluster: boolean;
-  dataSourceCallBackFunc: (dataSource: DataSourceOption) => void;
+  hideLocalCluster?: boolean;
+  dataSourceCallBackFunc?: (dataSource: DataSourceOption) => void;
   disableDataSourceSelectable?: boolean;
   className?: string;
   selectedOption?: DataSourceOption[];
   setMenuMountPoint?: (menuMount: MountPoint | undefined) => void;
+  filterFn?: (dataSource: any) => boolean;
 }
 
 export function DataSourceMenu(props: DataSourceMenuProps): ReactElement | null {
@@ -40,33 +43,41 @@ export function DataSourceMenu(props: DataSourceMenuProps): ReactElement | null 
     fullWidth,
     hideLocalCluster,
     selectedOption,
+    showDataSourceView,
+    filterFn,
   } = props;
 
-  if (!showDataSourceSelectable) {
+  if (!showDataSourceSelectable && !showDataSourceView) {
     return null;
   }
 
-  function renderMenu(className: string): ReactElement | null {
-    if (!showDataSourceSelectable) return null;
+  function renderDataSourceView(className: string): ReactElement | null {
+    if (!showDataSourceView) return null;
     return (
       <EuiHeaderLinks data-test-subj="top-nav" gutterSize="xs" className={className}>
-        {renderDataSourceSelectable()}
+        <DataSourceView
+          fullWidth={fullWidth}
+          selectedOption={selectedOption && selectedOption.length > 0 ? selectedOption : undefined}
+        />
       </EuiHeaderLinks>
     );
   }
 
-  function renderDataSourceSelectable(): ReactElement | null {
+  function renderDataSourceSelectable(className: string): ReactElement | null {
     if (!showDataSourceSelectable) return null;
     return (
-      <DataSourceSelectable
-        fullWidth={fullWidth}
-        hideLocalCluster={hideLocalCluster}
-        savedObjectsClient={savedObjects}
-        notifications={notifications.toasts}
-        onSelectedDataSource={dataSourceCallBackFunc}
-        disabled={disableDataSourceSelectable || false}
-        selectedOption={selectedOption && selectedOption.length > 0 ? selectedOption : undefined}
-      />
+      <EuiHeaderLinks data-test-subj="top-nav" gutterSize="xs" className={className}>
+        <DataSourceSelectable
+          fullWidth={fullWidth}
+          hideLocalCluster={hideLocalCluster || false}
+          savedObjectsClient={savedObjects!}
+          notifications={notifications!.toasts}
+          onSelectedDataSource={dataSourceCallBackFunc!}
+          disabled={disableDataSourceSelectable || false}
+          selectedOption={selectedOption && selectedOption.length > 0 ? selectedOption : undefined}
+          filterFn={filterFn}
+        />
+      </EuiHeaderLinks>
     );
   }
 
@@ -77,12 +88,18 @@ export function DataSourceMenu(props: DataSourceMenuProps): ReactElement | null 
       return (
         <>
           <MountPointPortal setMountPoint={setMenuMountPoint}>
-            {renderMenu(menuClassName)}
+            {renderDataSourceSelectable(menuClassName)}
+            {renderDataSourceView(menuClassName)}
           </MountPointPortal>
         </>
       );
     } else {
-      return <>{renderMenu(menuClassName)}</>;
+      return (
+        <>
+          {renderDataSourceSelectable(menuClassName)}
+          {renderDataSourceView(menuClassName)}
+        </>
+      );
     }
   }
 
@@ -91,4 +108,6 @@ export function DataSourceMenu(props: DataSourceMenuProps): ReactElement | null 
 
 DataSourceMenu.defaultProps = {
   disableDataSourceSelectable: false,
+  showDataSourceView: false,
+  showDataSourceSelectable: false,
 };